# Optimized large SIMD butterflies for better register usage