--- a/src/neon/neon_vector.rs
+++ b/src/neon/neon_vector.rs
@@ -147,6 +147,9 @@ pub trait NeonVector: Copy + Debug + Send + Sync {
     unsafe fn store_partial_lo_complex(ptr: *mut Complex<Self::ScalarType>, data: Self);
     unsafe fn store_partial_hi_complex(ptr: *mut Complex<Self::ScalarType>, data: Self);
 
+    // math ops
+    unsafe fn neg(a: Self) -> Self;
+
     /// Generates a chunk of twiddle factors starting at (X,Y) and incrementing X `COMPLEX_PER_VECTOR` times.
     /// The result will be [twiddle(x*y, len), twiddle((x+1)*y, len), twiddle((x+2)*y, len), ...] for as many complex numbers fit in a vector
     unsafe fn make_mixedradix_twiddle_chunk(
@@ -210,6 +213,11 @@ impl NeonVector for float32x4_t {
     unsafe fn store_partial_hi_complex(ptr: *mut Complex<Self::ScalarType>, data: Self) {
         let high = vget_high_f32(data);
         vst1_f32(ptr as *mut f32, high);
+    }
+
+    #[inline(always)]
+    unsafe fn neg(a: Self) -> Self {
+        vnegq_f32(a)
     }
 
     #[inline(always)]
@@ -313,6 +321,11 @@ impl NeonVector for float64x2_t {
     #[inline(always)]
     unsafe fn store_partial_hi_complex(_ptr: *mut Complex<Self::ScalarType>, _data: Self) {
         unimplemented!("Impossible to do a partial store of complex f64's");
+    }
+
+    #[inline(always)]
+    unsafe fn neg(a: Self) -> Self {
+        vnegq_f64(a)
     }
 
     #[inline(always)]