--- a/src/sse/sse_utils.rs
+++ b/src/sse/sse_utils.rs
@@ -63,7 +63,7 @@ impl Rotate90F32 {
         let temp = _mm_shuffle_ps(values, values, 0xB1);
         _mm_xor_ps(temp, self.sign_both)
     }
-    
+
     #[inline(always)]
     pub unsafe fn rotate_both_45(&self, values: __m128) -> __m128 {
         let rotated = self.rotate_both(values);