--- a/src/wasm_simd/wasm_simd_butterflies.rs
+++ b/src/wasm_simd/wasm_simd_butterflies.rs
@@ -2414,9 +2414,7 @@ pub struct WasmSimdF32Butterfly16<T> {
     bf4: WasmSimdF32Butterfly4<T>,
     twiddles_packed: [v128; 6],
     twiddle1: v128,
-    twiddle2: v128,
     twiddle3: v128,
-    twiddle6: v128,
     twiddle9: v128,
 }
 
@@ -2453,9 +2451,7 @@ impl<T: FftNum> WasmSimdF32Butterfly16<T> {
                     pack_32(tw6, tw9),
                 ],
                 twiddle1: pack_32(tw1, tw1),
-                twiddle2: pack_32(tw2, tw2),
                 twiddle3: pack_32(tw3, tw3),
-                twiddle6: pack_32(tw6, tw6),
                 twiddle9: pack_32(tw9, tw9),
             }
         }
@@ -2564,11 +2560,11 @@ impl<T: FftNum> WasmSimdF32Butterfly16<T> {
         let [in2, in3] = Self::load_parallel_chunk(&buffer, 2);
         let mut tmp2 = self.bf4.perform_parallel_fft_direct(in2);
         let mut tmp3 = self.bf4.perform_parallel_fft_direct(in3);
-        tmp2[1] = mul_complex_f32(tmp2[1], self.twiddle2);
+        tmp2[1] = self.bf4.rotate.rotate_both_45(tmp2[1]);
         tmp2[2] = self.bf4.rotate.rotate_both(tmp2[2]);
-        tmp2[3] = mul_complex_f32(tmp2[3], self.twiddle6);
+        tmp2[3] = self.bf4.rotate.rotate_both_135(tmp2[3]);
         tmp3[1] = mul_complex_f32(tmp3[1], self.twiddle3);
-        tmp3[2] = mul_complex_f32(tmp3[2], self.twiddle6);
+        tmp3[2] = self.bf4.rotate.rotate_both_135(tmp3[2]);
         tmp3[3] = mul_complex_f32(tmp3[3], self.twiddle9);
 
         // Do these last, because fewer twiddles means fewer temporaries forcing the above data to spill
@@ -2576,7 +2572,7 @@ impl<T: FftNum> WasmSimdF32Butterfly16<T> {
         let tmp0 = self.bf4.perform_parallel_fft_direct(in0);
         let mut tmp1 = self.bf4.perform_parallel_fft_direct(in1);
         tmp1[1] = mul_complex_f32(tmp1[1], self.twiddle1);
-        tmp1[2] = mul_complex_f32(tmp1[2], self.twiddle2);
+        tmp1[2] = self.bf4.rotate.rotate_both_45(tmp1[2]);
         tmp1[3] = mul_complex_f32(tmp1[3], self.twiddle3);
 
         // Size-4 FFTs down each pair of transposed columns, storing them as soon as we're done with them